+++ commands/dbaas/mongo/mongo.go
@@ -12,15 +12,16 @@ import (
 	"github.com/ionos-cloud/ionosctl/v6/commands/dbaas/mongo/user"
 	"github.com/ionos-cloud/ionosctl/v6/internal/core"
 	"github.com/spf13/cobra"
+	"golang.org/x/exp/slices"
 )
 
 func DBaaSMongoCmd() *core.Command {
-	nonDeprecatedAliases := []string{"mongodb", "mdb"}
+	deprecatedAliases := []string{"mongodb", "mg"}
 
 	mongoCmd := &core.Command{
 		Command: &cobra.Command{
 			Use:     "mongo",
-			Aliases: []string{"mongodb", "mdb", "m"},
+			Aliases: append(deprecatedAliases, "mdb", "m"),
 			Short:   "DBaaS Mongo Operations",
 			Long: `DBaaS Mongo Operations. Wiki: https://docs.ionos.com/cloud/managed-services/database-as-a-service/mongodb
 With IONOS Cloud Database as a Service (DBaaS) MongoDB, you can quickly set up and manage MongoDB database clusters. It is an open-source, NoSQL database solution that does not require a relational Database Management System (DBMS). The feature offers flexible data schemas, managed MongoDB solution with deployment and monitoring of your databases. To cater to your workload use cases, IONOS provides MongoDB editions such as Playground, Business, and Enterprise models.`,
@@ -30,10 +31,11 @@ With IONOS Cloud Database as a Service (DBaaS) MongoDB, you can quickly set up a
 
 	mongoCmd.Command.PersistentPreRun = func(cmd *cobra.Command, args []string) {
 		aliasUsed := os.Args[2]
-		if aliasUsed == "m" {
-			cmd.PrintErrf("WARNING: 'm' is deprecated and will be removed in a future release, "+
+		if slices.Contains(deprecatedAliases, aliasUsed) {
+			cmd.PrintErrf("WARNING: '%s' is deprecated and will be removed in a future release, "+
 				"please use any of '%s' instead.\n",
-				strings.Join(append([]string{mongoCmd.Command.Use}, nonDeprecatedAliases...), "', '"),
+				aliasUsed,
+				strings.Join(append([]string{mongoCmd.Command.Use}, deprecatedAliases...), "', '"),
 			)
 		}
 	}